Updated Player class
Added more functions to the player class. I'm getting this weird error when I terminate the bot, so I will need to fix that next time. Additionally, I need to change my functions and other features in the quizbot.py file so that it's not so similar to the tutorials I used to learn things about this API.