Changed files names and updated quizbot.py
Changed the name of the .env file to quizbot.py (to steer away from the format of the code that was in the tutorial I followed). Additionally, I updated the quizbot.py file to include my own code instead of tutorial code. New code has not yet been tested so the crucial parts of the old code have simply been commented out instead of removed. If this works, I will have shaved off some lines of code.