Organized files
Since I changed the way I import cogs, I switched up the organization of the cogs folder. Files in this folder that begin with a capital letter are NOT cogs, but are integral to the function of the cogs. I am reserving the underscore format to separate words in a file for variables within the files as well.