Major changes
WORK DONE:
- Filled out roundCreator.py
- Filled out playerActions.py
- Bug fixes and finalizing Player and Round classes

BUGS:
- Nested dictionary in roundCreator not working

TODO:
After I fix the nested dictionary issue in roundCreator, implement the same kind of dictionary in playerActions. Eventually these dictionaries will be moved to separate files so that data is kept even if the bot restarts (perhaps write a background task to update the file every so often).